mark QMLModule itself RUNTIME
this reintroduces behavior as before
26d111b771a8364cef25428a7e1831bf9f291ab8 where not finding QMLModule
(which happens when qmlplugindump is not present or cannot be executed) it
will be listed as missing runtime dependency. since qmldump is only used
to check for runtime deps it itself should be considered runtime in that
if it is missing all runtime checks cannot and will not be performed
correctly anyway.

REVIEW: 124913