label revision titles
Labels are appended to revisions that have been accepted but are still
open and those that are waiting for revision. Unicode glyphs are used,
a chechkmark for accepted (as on the website), a ballot cross for
revisions that require action to proceed (the Phab icon doesn't seem to
have a Unicode glyph equivalent; the more appropriate "Raised Hand" glyph
is less common in fonts and less readable in usual GUI font sizes).