Add API to control the disabled plugins
Summary:
We already had a way to control that list from the system settings
thanks to previous changes. Now this is also exposed via API.
Note however that the blacklist coming from the settings can't be
ignored, we control via the API only an extra blacklist sitting on top.

Also, on desktop platforms, the blacklist contains "saveasplugin" by
default.

Reviewers: aacid, apol

Differential Revision: https://phabricator.kde.org/D8341