Add Id to Twitter plugin
KPluginMetaData infers plugin Id from the plugin file name, which works for
binary (.so) plugins, but does not work for JS-only plugins like the Twitter
plugin where the plugin file is called main.js. In order to be able to
blacklist the Twitter plugin we need to give it a proper Id.

Differential Revision: https://phabricator.kde.org/D9203