Set Dialog flag for JobDialog
Summary:
The JobDialog window is treated as a "normal" window. When using a tiling window manager it gets as much space assigned as the other windows, which is quite unnecessary. With this patch it has the same size as when using a
floating WM and is floating above the "normal" windows (e.g. Dolphin), which seems more appropriate.

Test Plan:
Before:
{F6161449}

After:
{F6161451}

Reviewers: apol, #frameworks, cfeck

Reviewed By: apol

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D14472