[email] Fix lauching Thunderbird from Flatpak
Previously it wouldn't detect that Thunderbird is installed as a Flatpak and even if it did it would fail since Flatpak needs special care to pass the url into it.

Fortunately the special stuff is nicely handled in KService/KIO. Unfortunatley this means we cannot use -compose to set body/subject.

There is an issue in the Thunderbird desktop file that prevents it from working correctly: https://github.com/flathub/org.mozilla.Thunderbird/pull/124

BUG: 424383