Rework job view handling
The job view handling is a large chunk of quite imperative QML code that is hard to maintain. It also seems to suffer from QML/C++ ownership issues that result in crashes.

This patch introduces a new JobView component that does the job visualization. Internally it uses a new JobController C++ class that is responsible for tracking the job state.

JobDialog and AlternativesView are ported to use it.

Visually the end result is similar, except for the running job aspect. Instead of the de-facto not working progress bar a BusyIndicator is used.

BUG: 419170