When setting QML2_IMPORT_PATH on tests, preserve original value too
Currently autotests are setting the QML2_IMPORT_PATH environment variable
to point at the binary dir so that Purpose's QML files can be found.
However this overwrites the existing env var, which may be already set
when running the tests. In KDE's CI, this means the tests can't find
Kirigami, which is installed in a non-standard path.

I'm changing the CMake code so that the binary dir is prepended to the
existing value of the env var, and using the correct separator so that it
(hopefully?) works on Windows too.