Evaluate initial configuration
The initial configuration may be valid already, but it is only evaluated
once something changes.

The nextcloud plugin currently works around that. This patch will allow
to drop this workaround.

For kdeconnect and bluetooth change the initial values from empty string
to undefined to prevent the empty string being evaluated as a set value