[imgur] Copy link to clipboard and show notification
Currently it is up to the application to present the result url to the user, and they do that quite inconsistently.

Some apps open the url, some show it in an inline message, some copy it to the clipboard.

With this patch the link is always copied to the clipboard and a notification informs the user that this has been done.

This also removes the url property from the output which automatically removes the applications's own message handling

BUG: 437347