[kdeconnect] Port away from org.kde.kdeconnect QML import
The code assumes that the KDE Connect QML module can be imported.

This isn't necessarily the case, e.g. because the module was built against
a differe Qt version or because we are in a Flatpak.

Since we are only interested in listing connected devices in a model do that
outselves. It's not much code anyway