Clipboard: Set clipboard text, too, if applicable
Purpose has no proper API contract for sending just text, it's all sent
as part of "urls".

Try to guess what is an URL (we want to be tolerant in the parser since
we don't really know how strict the input will be) and only set that as URL
and also set everything as text.

CCBUG: 501811